fix(nwc): support wallets without a kind-13194 info event + back off on relay ban (#737)
CoinOS (and any NIP-47 wallet that doesn't publish a kind-13194 info event)
failed every operation with "no info event (kind 13194) returned from relay":
@getalby/sdk negotiates encryption from that event before every command and
throws when it's absent. Confirmed via direct relay query — CoinOS publishes
zero kind-13194 events yet actively serves kind-23195 NWC responses (NIP-04
ciphertext), so the wallet works; we just refused to talk to it.

- src/services/nwcEncryption.ts (new): after connecting, probe once for the
  info event; if missing, pin the client to NIP-04 (the SDK's own baseline).
  The per-wallet decision is cached so reconnects skip the probe.
- Treat relay rejections ("temp-banned" / rate-limit) as connection errors and
  feed fire-and-forget publish failures into the cooldown, so a rate-limiting
  relay gets parked fast instead of looping publish→ban→reconnect→publish.
- src/services/nwcErrors.ts (new): extracted the NWC error-classification +
  abort/cancellation primitives so nwcService drops back under the 1000-line
  cap (1024 → 944). Public API preserved via re-export.

Closes #737
